+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.1.1
+* Update README
+* Update example
+
+## 0.1.0
+
+* `RadioOption` accepts any object as it's value
+* `RadioOption` parameter `text` is renamed to `label`
+* Updated example
+
+
 ## 0.0.2
 
 * Improved documentation and added example